Add support for dynamic pclk switching (and PLL reconfig)
Routines to initialise/lock the pclk PLL, and bitbang test_mode config
words into it at runtime.

This is complemented by a cleanup of the line-doubling code, which tries a few different
pixel clocks in order to find a good fit for a given mode.

(With this, one can arbitrarily change into HR mode, wide modes like 16, etc.)